requirements.txt
supabase==2.3.4
python-dotenv==1.0.0
pyyaml==6.0.1
